pyproject.toml
[project]
name = "aegis"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-groq>=0.1.0",
    "langchain-core>=0.2.0",
    "langsmith>=0.1.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "python-docx>=1.1.0",
    "redis>=5.0.0",
    "arq>=0.26.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
